@@ -3,19 +3,26 @@ module.exports = {
     browser: true,
     es6: true,
     node: true,
+    jest: true,
   },
-  extends: 'airbnb-base',
+  extends: ['eslint:recommended',
+            'plugin:vue/recommended'],
   globals: {
     Atomics: 'readonly',
     SharedArrayBuffer: 'readonly',
   },
+  parser: 'vue-eslint-parser',
   parserOptions: {
+    parser: 'babel-eslint',
     ecmaVersion: 2019,
     sourceType: 'module',
+    allowImportExportEverywhere: true
   },
   plugins: [
-    'vue'
+    'vue',
+    'import'
   ],
   rules: {
-  },
+    'vue/no-unused-vars': 'error'
+  }
 };
